4/26/2020 review
as per this line in exercise requirements: Large Display: The underlying grid structure should be identical- tried to emulate original design for grid and image placement.
chose and sized images all 2X3 aspect ratio and exported for retina display
placed two images in html
TODO: green pattern image is not responding well and is not taking the full size of the parent column.
finalizing for foundation